cluster/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
numpy
torch
torchvision
lightning
matplotlib
pandas
diffusers
transformers
accelerate
pillow